requirements.txt
altgraph==0.17.2
appdirs==1.4.4
APScheduler==3.8.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
distlib==0.3.3
Django==3.2.9
django-recaptcha==3.0.0
emoji==1.6.1
filelock==3.4.0
future==0.18.2
gunicorn==20.1.0
keyboard==0.13.5
Markdown==3.3.6
MouseInfo==0.1.3
numpy==1.22.3
opencv-python==4.5.5.64
pefile==2021.9.3
Pillow==9.1.0
platformdirs==2.4.0
PyAutoGUI==0.9.53
PyDirectInput==1.0.4
PyGetWindow==0.0.9
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-decouple==3.5
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywin32-ctypes==0.2.0
screeninfo==0.8
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
virtualenv==20.10.0
whitenoise==5.3.0
